Refactoring + listener
I refactored a little and I added the listener for the submit button. I
also written a method to check if all questions are answered but didn't
set it to a listener. Problem is now that if you answer to only a part
of the questions, you will get only those answers.
PS: the syso from the listener can be removed, I left it to be easy to
see the result.